requirements.txt
Django==3.2
djangorestframework==3.13.1
gunicorn==20.1.0
drf-yasg==1.20.0
django-environ==0.8.1
psycopg2-binary==2.8.6
